[DDW-685] Remove duplicated code
