+++ source/renderer/app/components/staking/stake-pools/StakePools.scss
@@ -1,0 +1,17 @@
+@import '../stakingConfig';
+
+.component {
+  @extend %regularText;
+  margin: 20px;
+
+  h2 {
+    color: var(--theme-staking-stake-pools-title-color);
+    line-height: 1.38;
+    margin-bottom: 10px;
+    opacity: 0.5;
+    padding-left: 20px;
+  }
+}
+
+.poolsListsContainer {
+}
