+++ source/renderer/app/components/staking/stake-pools/StakePoolsSearch.scss
@@ -1,0 +1,47 @@
+.component {
+  margin-bottom: 26px;
+  position: relative;
+}
+
+.searchIcon {
+  left: 20px;
+  position: absolute;
+  top: 13.5px;
+  z-index: 1;
+
+  svg {
+    width: 15px;
+  }
+}
+
+.searchInput {
+  input {
+    border-radius: 4px;
+    height: 50px;
+    padding-left: 50px;
+    padding-right: 175px;
+  }
+}
+
+.searchFilter {
+  display: block;
+  position: absolute;
+  right: 20px;
+  top: 13px;
+
+  li {
+    display: inline-block;
+    margin-left: 20px;
+  }
+
+  button {
+    color: var(--theme-staking-stake-pools-search-button-color);
+    cursor: pointer;
+    font-size: 16px;
+    opacity: 0.3;
+  }
+
+  .searchFilterActiveItem {
+    opacity: 1;
+  }
+}
