+++ storybook/stories/Staking.stories.js
@@ -15,7 +15,7 @@ import DelegationCenter from '../../source/renderer/app/components/staking/deleg
 import StakingEpochs from '../../source/renderer/app/components/staking/epochs/StakingEpochs';
 import StakingInfo from '../../source/renderer/app/components/staking/info/StakingInfo';
 
-import { StakePoolsListStory } from './StakePoolsListStory.js';
+import { StakePoolsStory } from './StakePoolsStory.js';
 import { StakingRewardsStory } from './Staking-Rewards.stories';
 
 const defaultPercentage = 10;
@@ -95,7 +95,7 @@ storiesOf('Staking', module)
     { id: 'delegation-center' }
   )
 
-  .add(pageNames['stake-pools'], StakePoolsListStory, { id: 'stake-pools' })
+  .add(pageNames['stake-pools'], StakePoolsStory, { id: 'stake-pools' })
 
   .add(pageNames.rewards, StakingRewardsStory, { id: 'rewards' })
